Cloud_Function/Data_processing_cloud/requirements.txt
numpy==1.23.5
pandas==1.5.3
jsonpointer==2.1
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
gcloud==0.18.3
cloudpathlib==0.18.1
aiosqlite==0.19.0
psycopg2==2.9.9
google-cloud-secret-manager
google-cloud-storage
psycopg2-binary
google-cloud-logging
google-cloud-aiplatform
google-cloud-pubsub
pytz
openpyxl
fuzzywuzzy
python-Levenshtein